Added timestamp to file name
Each image will have the date/time YYMMDD_HHMMSS in the file name so that they won't be overwritten when the pi's reboot. I will synchronize the pi's before I begin the work so that their clocks match (even though it won't be the right time after a few reboots, it should still match).